requirements.txt
dash==4.0.0
dash_auth_external==1.2.1
dash_bootstrap_components==2.0.4
jsonschema==4.26.0
matplotlib==3.8.4
more_itertools==11.0.2
numpy==1.23.5
pandas==1.5.3
plotly==5.9.0
python-dotenv==1.2.2
python_dateutil==2.8.2
reportlab==3.6.13
Requests==2.33.1
scipy==1.10.0
websocket_client==0.58.0
